pyproject.toml
[project]
name = "politiscales-ai"
version = "1.2.0"
description = "Benchmarking AI models on the PolitiScales political test."
authors = [
    { name = "PolitiScales-AI Contributors" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["runner"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88
